Improve boss teleporting and gating

+++ src/Inferno.Core/Robot.h
@@ -87,7 +87,7 @@ namespace Inferno {
         // Joint lookup for each gun and animation state
         JointLookup Joints[MAX_GUNS + 1][N_ANIM_STATES];
 
-        List<int> GatedRobots; // Robots to gate in when hit. For bosses.
+        List<int8> GatedRobots; // Robots to gate in when hit. For bosses.
         float TeleportInterval; // Interval between boss teleports.
     };
